modify category def & add poorResult desc field

+++ src/tw/jms/loyal/DbAdmin.java
@@ -1,19 +1,29 @@
 package tw.jms.loyal;
 
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+
 import tw.jms.loyal.dao.DaoSqliteImpl;
 import tw.jms.loyal.dao.ElasticSearchDao;
 import tw.jms.loyal.utils.IndexConstants;
 
 public class DbAdmin {
 
-	public static void main(String args[]) {
+	public static void main(String args[]) throws IOException {
 		if (args.length == 0) {
 			printUsage();
 		} else if (args[0].equals("initialize")) {
 			ElasticSearchDao.createIndex(IndexConstants.INDEX_PROVIDENCE,
 					IndexConstants.TYPE_WORD);
 		} else if (args[0].equals("destroy")) {
-			ElasticSearchDao.deleteIndex(IndexConstants.INDEX_PROVIDENCE);
+			System.out.print("Do you really want to destroy ElasticSearch DB? :");
+			BufferedReader in = new BufferedReader(new InputStreamReader(System.in));
+			String s = null;
+			if((s = in.readLine()) != null && s.equals("yes")){	
+				System.out.println("deleting index");
+				ElasticSearchDao.deleteIndex(IndexConstants.INDEX_PROVIDENCE);
+			}
 		} else if (args[0].equals("init_user_db")) {
 			DaoSqliteImpl dao = new DaoSqliteImpl();
 			dao.initialize();
